requirements.txt
#openai
pydub
openai-whisper